Indirect Left Recursion Detection
Each assignment is passed to an annotator. If the assignment is left recursive then the annotator highlights the code for the that assignment.

BUG FIXED:
Annotator would get stuck in an infinite loop when a non-recursive assignment pointed to a recursive assignment.
Annotator now knows which assignments it has looked at to prevent it getting stuck in a loop.